requirements.txt
streamlit
openai
pillow
transformers
gpt4all
opencv-python 
accelerate 
insightface 
onnx == 1.16.1
onnxruntime
ultralytics
torch
diffusers